Moves post page to components/pages
