+++ pages/post/[id].js
@@ -1,21 +1,4 @@
 import React from "react";
-import { useRouter } from "next/router";
-import { Main } from "core/Main";
-import { PageHeader } from "composed/PageHeader";
-import { posts } from "posts";
+import { Post } from "pages/post";
 
-const Post = () => {
-  const {
-    query: { id }
-  } = useRouter();
-  const blogPost = posts[id];
-
-  return (
-    <>
-      <PageHeader />
-      <Main>{blogPost && <blogPost.Component />}</Main>
-    </>
-  );
-};
-
-export default Post;
+export default () => <Post />;
